+++ tests/reduct/bucket_api_test.cc
@@ -2,54 +2,54 @@
 
 #include <catch2/catch.hpp>
 
-#include "helpers.h"
+#include "fixture.h"
 #include "reduct/client.h"
 
 using reduct::Error;
 using reduct::IBucket;
 using reduct::IClient;
 
+using s = std::chrono::seconds;
+
+constexpr auto kBucketName = "bucket";
+
 TEST_CASE("reduct::Client should create a bucket", "[bucket_api]") {
-  auto client = BuildClient();
-  const auto kBucketName = RandomBucketName();
-  auto [bucket, err] = client->CreateBucket(kBucketName);
+  Fixture ctx;
+  auto [bucket, err] = ctx.client->CreateBucket(kBucketName);
 
   REQUIRE(err == Error::kOk);
   REQUIRE(bucket);
 
   SECTION("should return HTTP status") {
-    auto [null, err_409] = client->CreateBucket(kBucketName);
+    auto [null, err_409] = ctx.client->CreateBucket(kBucketName);
     REQUIRE(err_409 == Error{.code = 409, .message = fmt::format("Bucket '{}' already exists", kBucketName)});
     REQUIRE_FALSE(null);
   }
 }
 
 TEST_CASE("reduct::Client should get a bucket", "[bucket_api]") {
-  auto client = BuildClient();
-  const auto kBucketName = RandomBucketName();
-  [[maybe_unused]] auto _ = client->CreateBucket(kBucketName);
-
-  auto [bucket, err] = client->GetBucket(kBucketName);
+  Fixture ctx;
+  [[maybe_unused]] auto _ = ctx.client->CreateBucket(kBucketName);
+  auto [bucket, err] = ctx.client->GetBucket(kBucketName);
 
   REQUIRE(err == Error::kOk);
   REQUIRE(bucket);
 
   SECTION("should return HTTP status") {
-    auto [null, err_404] = client->GetBucket("XXXXXX");
+    auto [null, err_404] = ctx.client->GetBucket("XXXXXX");
     REQUIRE(err_404 == Error{.code = 404, .message = "Bucket 'XXXXXX' is not found"});
     REQUIRE_FALSE(null);
   }
 }
 
 TEST_CASE("reduct::IBucket should have settings", "[bucket_api]") {
-  auto client = BuildClient();
-  const auto kBucketName = RandomBucketName();
+  Fixture ctx;
   const IBucket::Settings kSettings{
       .max_block_size = 100,
       .quota_type = IBucket::QuotaType::kFifo,
       .quota_size = 1000,
   };
-  auto [bucket, create_err] = client->CreateBucket(kBucketName, kSettings);
+  auto [bucket, create_err] = ctx.client->CreateBucket(kBucketName, kSettings);
 
   REQUIRE(create_err == Error::kOk);
   REQUIRE(bucket);
@@ -97,10 +97,8 @@ TEST_CASE("reduct::IBucket should have settings", "[bucket_api]") {
 }
 
 TEST_CASE("reduct::IBucket should get bucket stats", "[bucket_api]") {
-  auto client = BuildClient();
-
-  const auto kBucketName = RandomBucketName();
-  auto [bucket, _] = client->CreateBucket(kBucketName);
+  Fixture ctx;
+  auto [bucket, _] = ctx.client->CreateBucket(kBucketName);
 
   auto t = IBucket::Time();
   REQUIRE(bucket->Write("entry-1", "some_data", t) == Error::kOk);
@@ -119,42 +117,35 @@ TEST_CASE("reduct::IBucket should get bucket stats", "[bucket_api]") {
 }
 
 TEST_CASE("reduct::IBucket should get list of entries", "[bucket_api]") {
-  auto client = BuildClient();
+  Fixture ctx;
 
-  const auto kBucketName = RandomBucketName();
-  auto [bucket, _] = client->CreateBucket(kBucketName);
-
-  auto t = IBucket::Time();
-  REQUIRE(bucket->Write("entry-1", "some_data", t) == Error::kOk);
-  REQUIRE(bucket->Write("entry-2", "some_data", t + std::chrono::microseconds(100)) == Error::kOk);
-
-  auto [entries, err] = bucket->GetEntryList();
+  auto [entries, err] = ctx.bucket_1->GetEntryList();
   REQUIRE(err == Error::kOk);
   REQUIRE(entries.size() == 2);
+
+  auto t = IBucket::Time();
   REQUIRE(entries[0] == IBucket::EntryInfo{
                             .name = "entry-1",
-                            .record_count = 1,
+                            .record_count = 2,
                             .block_count = 1,
-                            .size = 9,
-                            .oldest_record = t,
-                            .latest_record = t,
+                            .size = 12,
+                            .oldest_record = t + s(1),
+                            .latest_record = t + s(2),
                         });
 
   REQUIRE(entries[1] == IBucket::EntryInfo{
                             .name = "entry-2",
-                            .record_count = 1,
+                            .record_count = 2,
                             .block_count = 1,
-                            .size = 9,
-                            .oldest_record = t + std::chrono::microseconds(100),
-                            .latest_record = t + std::chrono::microseconds(100),
+                            .size = 12,
+                            .oldest_record = t + s(3),
+                            .latest_record = t + s(4),
                         });
 }
 
 TEST_CASE("reduct::IBucket should remove a bucket", "[bucket_api]") {
-  auto client = BuildClient();
-
-  const auto kBucketName = RandomBucketName();
-  auto [bucket, _] = client->CreateBucket(kBucketName);
+  Fixture ctx;
+  auto [bucket, _] = ctx.client->CreateBucket(kBucketName);
 
   REQUIRE(bucket->Remove() == Error::kOk);
   REQUIRE(bucket->Remove() == Error{.code = 404, .message = fmt::format("Bucket '{}' is not found", kBucketName)});